+++ packages/playwright-test/src/matchers/matchers.ts
@@ -270,6 +270,7 @@ export function toHaveURL(
   }, expected, options);
 }
 
+
 export async function toBeOK(
   this: ReturnType<Expect['getState']>,
   response: APIResponseEx
