+++ packages/playwright-test/src/matchers/toMatchSnapshot.ts
@@ -14,10 +14,16 @@
  * limitations under the License.
  */
 
+/* eslint-disable no-console */
 import type { Expect } from '../types';
 import { currentTestInfo } from '../globals';
-import { compare } from './golden';
-import { addSuffixToFilePath, sanitizeForFilePath, trimLongString } from '../util';
+import { mimeTypeToComparator, ComparatorResult, ImageComparatorOptions } from './comparators';
+import { addSuffixToFilePath, serializeError, sanitizeForFilePath, trimLongString } from '../util';
+import { UpdateSnapshots } from '../types';
+import colors from 'colors/safe';
+import fs from 'fs';
+import path from 'path';
+import { TestInfoImpl } from '../testInfo';
 
 // from expect/build/types
 type SyncExpectationResult = {
@@ -25,13 +31,27 @@ type SyncExpectationResult = {
   message: () => string;
 };
 
+const extensionToMimeType: { [key: string]: string } = {
+  'dat': 'application/octet-string',
+  'jpeg': 'image/jpeg',
+  'jpg': 'image/jpeg',
+  'png': 'image/png',
+  'txt': 'text/plain',
+};
+
+
 type NameOrSegments = string | string[];
 const SNAPSHOT_COUNTER = Symbol('noname-snapshot-counter');
-export function toMatchSnapshot(this: ReturnType<Expect['getState']>, received: Buffer | string, nameOrOptions: NameOrSegments | { name: NameOrSegments, threshold?: number }, optOptions: { threshold?: number, pixelCount?: number, pixelRatio?: number } = {}): SyncExpectationResult {
-  let options: { name: NameOrSegments, threshold?: number, pixelCount?: number, pixelRatio?: number };
-  const testInfo = currentTestInfo();
-  if (!testInfo)
-    throw new Error(`toMatchSnapshot() must be called during the test`);
+
+type MatchSnapshotOptions = { threshold?: number, pixelCount?: number, pixelRatio?: number };
+
+function parseMatchSnapshotOptions(
+  testInfo: TestInfoImpl,
+  snapshotExtension: string,
+  nameOrOptions: NameOrSegments | { name: NameOrSegments } & MatchSnapshotOptions,
+  optOptions: MatchSnapshotOptions = {},
+) {
+  let options: { name: NameOrSegments } & ImageComparatorOptions;
   if (Array.isArray(nameOrOptions) || typeof nameOrOptions === 'string')
     options = { name: nameOrOptions, ...optOptions };
   else
@@ -42,7 +62,7 @@ export function toMatchSnapshot(this: ReturnType<Expect['getState']>, received: 
       ...testInfo.titlePath.slice(1),
       (testInfo as any)[SNAPSHOT_COUNTER],
     ].join(' ');
-    options.name = sanitizeForFilePath(trimLongString(fullTitleWithoutSpec)) + determineFileExtension(received);
+    options.name = sanitizeForFilePath(trimLongString(fullTitleWithoutSpec)) + '.' + snapshotExtension;
   }
 
   options = {
@@ -58,37 +78,187 @@ export function toMatchSnapshot(this: ReturnType<Expect['getState']>, received: 
 
   // sanitizes path if string
   const pathSegments = Array.isArray(options.name) ? options.name : [addSuffixToFilePath(options.name, '', undefined, true)];
-  const withNegateComparison = this.isNot;
+  const snapshotPath = testInfo.snapshotPath(...pathSegments);
+  const outputFile = testInfo.outputPath(...pathSegments);
+  const expectedPath = addSuffixToFilePath(outputFile, '-expected');
+  const actualPath = addSuffixToFilePath(outputFile, '-actual');
+  const diffPath = addSuffixToFilePath(outputFile, '-diff');
+
   let updateSnapshots = testInfo.config.updateSnapshots;
   if (updateSnapshots === 'missing' && testInfo.retry < testInfo.project.retries)
     updateSnapshots = 'none';
-  const { pass, message, expectedPath, actualPath, diffPath, mimeType } = compare(
-      received,
-      pathSegments,
+  const mimeType = extensionToMimeType[snapshotExtension] ?? 'application/octet-string';
+  const comparator = mimeTypeToComparator[mimeType];
+  if (!comparator)
+    throw new Error('Failed to find comparator with type ' + mimeType + ': ' + snapshotPath);
+  return {
+    snapshotPath,
+    hasSnapshotFile: fs.existsSync(snapshotPath),
+    expectedPath,
+    actualPath,
+    diffPath,
+    comparator,
+    mimeType,
+    updateSnapshots,
+    options,
+  };
+}
+
+export function toMatchSnapshot(
+  this: ReturnType<Expect['getState']>,
+  received: Buffer | string,
+  nameOrOptions: NameOrSegments | { name: NameOrSegments } & MatchSnapshotOptions,
+  optOptions: MatchSnapshotOptions = {}
+): SyncExpectationResult {
+  const testInfo = currentTestInfo();
+  if (!testInfo)
+    throw new Error(`toMatchSnapshot() must be called during the test`);
+  const {
+    options,
+    updateSnapshots,
+    snapshotPath,
+    hasSnapshotFile,
+    expectedPath,
+    actualPath,
+    diffPath,
+    mimeType,
+    comparator,
+  } = parseMatchSnapshotOptions(testInfo, determineFileExtension(received), nameOrOptions, optOptions);
+  if (!hasSnapshotFile)
+    return commitMissingSnapshot(testInfo, received, snapshotPath, actualPath, updateSnapshots, this.isNot);
+  const expected = fs.readFileSync(snapshotPath);
+  const result = comparator(received, expected, options);
+  return commitComparatorResult(
       testInfo,
+      expected,
+      received,
+      result,
+      mimeType,
+      snapshotPath,
+      expectedPath,
+      actualPath,
+      diffPath,
       updateSnapshots,
-      withNegateComparison,
-      options
+      this.isNot,
   );
-  const contentType = mimeType || 'application/octet-stream';
-  if (expectedPath)
-    testInfo.attachments.push({ name: 'expected', contentType, path: expectedPath });
-  if (actualPath)
-    testInfo.attachments.push({ name: 'actual', contentType, path: actualPath });
-  if (diffPath)
-    testInfo.attachments.push({ name: 'diff', contentType, path: diffPath });
-  return { pass, message: () => message || '' };
+}
+
+function commitMissingSnapshot(
+  testInfo: TestInfoImpl,
+  actual: Buffer | string,
+  snapshotPath: string,
+  actualPath: string,
+  updateSnapshots: UpdateSnapshots,
+  withNegateComparison: boolean,
+) {
+  const isWriteMissingMode = updateSnapshots === 'all' || updateSnapshots === 'missing';
+  const commonMissingSnapshotMessage = `${snapshotPath} is missing in snapshots`;
+  if (withNegateComparison) {
+    const message = `${commonMissingSnapshotMessage}${isWriteMissingMode ? ', matchers using ".not" won\'t write them automatically.' : '.'}`;
+    return { pass: true , message: () => message };
+  }
+  if (isWriteMissingMode) {
+    fs.mkdirSync(path.dirname(snapshotPath), { recursive: true });
+    fs.mkdirSync(path.dirname(actualPath), { recursive: true });
+    fs.writeFileSync(snapshotPath, actual);
+    fs.writeFileSync(actualPath, actual);
+  }
+  const message = `${commonMissingSnapshotMessage}${isWriteMissingMode ? ', writing actual.' : '.'}`;
+  if (updateSnapshots === 'all') {
+    console.log(message);
+    return { pass: true, message: () => message };
+  }
+  if (updateSnapshots === 'missing') {
+    testInfo._failWithError(serializeError(new Error(message)), false /* isHardError */);
+    return { pass: true, message: () => '' };
+  }
+  return { pass: false, message: () => message };
+}
+
+function commitComparatorResult(
+  testInfo: TestInfoImpl,
+  expected: Buffer | string,
+  actual: Buffer | string,
+  result: ComparatorResult,
+  mimeType: string,
+  snapshotPath: string,
+  expectedPath: string,
+  actualPath: string,
+  diffPath: string,
+  updateSnapshots: UpdateSnapshots,
+  withNegateComparison: boolean,
+) {
+  if (!result) {
+    if (withNegateComparison) {
+      const message = [
+        colors.red('Snapshot comparison failed:'),
+        '',
+        indent('Expected result should be different from the actual one.', '  '),
+      ].join('\n');
+      return {
+        pass: true, message: () => message,
+      };
+    }
+    return { pass: true, message: () => '' };
+  }
+
+  if (withNegateComparison)
+    return { pass: false, message: () => '' };
+
+  if (updateSnapshots === 'all') {
+    fs.mkdirSync(path.dirname(snapshotPath), { recursive: true });
+    fs.writeFileSync(snapshotPath, actual);
+    console.log(snapshotPath + ' does not match, writing actual.');
+    return {
+      pass: true,
+      message: () => snapshotPath + ' running with --update-snapshots, writing actual.'
+    };
+  }
+
+  fs.mkdirSync(path.dirname(expectedPath), { recursive: true });
+  fs.mkdirSync(path.dirname(actualPath), { recursive: true });
+  fs.writeFileSync(expectedPath, expected);
+  fs.writeFileSync(actualPath, actual);
+  if (result.diff)
+    fs.writeFileSync(diffPath, result.diff);
+
+  const output = [
+    colors.red(`Snapshot comparison failed:`),
+  ];
+  if (result.errorMessage) {
+    output.push('');
+    output.push(indent(result.errorMessage, '  '));
+  }
+  output.push('');
+  output.push(`Expected: ${colors.yellow(expectedPath)}`);
+  output.push(`Received: ${colors.yellow(actualPath)}`);
+  if (result.diff)
+    output.push(`    Diff: ${colors.yellow(diffPath)}`);
+
+  testInfo.attachments.push({ name: 'expected', contentType: mimeType, path: expectedPath });
+  testInfo.attachments.push({ name: 'actual', contentType: mimeType, path: actualPath });
+  if (result.diff)
+    testInfo.attachments.push({ name: 'diff', contentType: mimeType, path: diffPath });
+
+  return {
+    pass: false,
+    message: () => output.join('\n'),
+  };
+}
+
+function indent(lines: string, tab: string) {
+  return lines.replace(/^(?=.+$)/gm, tab);
 }
 
 
 function determineFileExtension(file: string | Buffer): string {
   if (typeof file === 'string')
-    return '.txt';
+    return 'txt';
   if (compareMagicBytes(file, [0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a]))
-    return '.png';
+    return 'png';
   if (compareMagicBytes(file, [0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10, 0x4a, 0x46, 0x49, 0x46, 0x00, 0x01, 0x01]))
-    return '.jpg';
-  return '.bin';
+    return 'jpg';
+  return 'dat';
 }
 
 function compareMagicBytes(file: Buffer, magicBytes: number[]): boolean {
